fix(hyperdim): resolve CI failures and maintain LOC gate
- Fixed AArch64/NEON linting errors by removing unnecessary `unsafe` blocks and correctly wrapping pointer/intrinsic operations.
- Resolved unreachable code and unused import warnings in `src/hyperdim.rs`.
- Extracted `tests` module from `src/hyperdim.rs` to `src/hyperdim_tests.rs` to comply with the 500 LOC per file limit.
- Maintained performance improvements for `cosine_similarity` (~79%) and `batch_cosine_similarity` (~41%).
